Removed second activity
I removed the second activity page I randomly had. It was not necessary to have. I have the home menu with the hamburger button and added a "Add Device" button that goes to a user input page. The adding functionality does not work yet. Still trying to get that to show up on the Recycler View list